feat: 모범 로드맵 생성 및 추천 기능 개발 완료

+++ src/main/java/com/consoltant/consoltant/domain/bestroadmap/entity/BestRoadmap.java
@@ -31,7 +31,7 @@ public class BestRoadmap {
     private User user;
 
     @Column(nullable = false)
-    private Long salary;
+    private Integer salary;
 
     @Column(nullable = false)
     private Long startAsset;
